refactor(webapp): route remaining L0 primitives through leaf cn (A1 hygiene)
Flip the 7 app-only components/ui primitives (carousel, dialog, select, sheet,
sonner, table, toggle) from importing cn via @/lib/utils to the leaf @/lib/cn,
matching the rest of the unified L0 layer. No behavior change (same cn function).

Makes the entire components/ui layer coupling-proof by default: if any of these is
ever shimmed from the widget side later, it can't re-trigger the @/lib/utils ->
wagmi-config (createConfig) import chain that broke widget tests during A1.

Verified: tsc --noEmit green; eslint clean (0 errors) on the changed files.
