+++ apps/webapp/src/components/ui/dialog.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 
 const Dialog = DialogPrimitive.Root;
