+++ apps/webapp/src/components/ui/table.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 import { HTMLMotionProps, motion } from 'motion/react';
 import { fadeAnimations } from '@/modules/ui/animation/presets';
